requirements.txt
llama-index
langchain
langchain-openai
python-dotenv